pyproject.toml
[tool.poetry]
name = "orders_server"
version = "0.1.0"
description = "Orders server"
authors = ["Vladimir Tsiuman <vladimir.tsyuman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4"

[tool.poetry.group.test.dependencies]
waiting = "^1.4.1"
faker = "^18.11.1"
pytest = "^7.4.0"
assertpy = "^1.1"
requests = "^2.31.0"
allure-pytest = "^2.13.2"
pydantic = "^1.10.9"


[tool.poetry.group.development.dependencies]
Flask = "2.2.1"
SQLAlchemy = "1.4.43"
marshmallow = "3.13.0"
webargs = "8.2.0"
typing = "3.7.4.3"
psycopg2-binary = "2.9.3"
alembic = "1.8.1"
pydantic = "^1.10.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
